+++ src/Components/index.js
@@ -1,0 +1,2 @@
+export { default as Loading } from "./Loading";
+export { default as PrivateRouter } from "./PrivateRouter";
