+++ src/Pages/index.js
@@ -1,0 +1,2 @@
+export { default as Login } from "./Login";
+export { default as Todos } from "./Todos";
